feat(UI): implemented undo button
